conversion to ts => command/events loader + typing
