+++ tsconfig.json
@@ -35,6 +35,7 @@
       "@commands/*": ["./commands/*"],
       "@events/*": ["./events/*"],
       "@config/*": ["./config/*"],
+      "@typing": ["./src/typing/index.ts"],
     },                                                   /* Specify a set of entries that re-map imports to additional lookup locations. */
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
